Added Bear and Tried To Figure Out Display Properties
Adam added the bear today and defined its class and id to include the
sprite sheet of the bear. The bear performs the directional movement
functions but the display properties still need to be messed with to get
the animation to time up with the duration of the movement function.